requirements.txt
django==5.0.6
psycopg2-binary==2.9.9
pillow==10.3.0
django-tables2==2.7.0
django-filter==24.2
django-select2==8.1.2
django-forms-dynamic==1.0.0
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
django-ninja==1.1.0
django-ninja-extra==0.20.7
requests==2.32.3
